refactor(molecules/CollageCard): 컴포넌트 분리
